Check that environment suites have no clashing IDs
A user can pass lists of different environment suites to run, since
environments with different interfaces might be in different registries,
but agents might have compatibility with more than one env type. Within
a registry, environments are guaranteed to not have clashing IDs, but we
add a check here to make sure that there are also no clashing IDs
between the different registries.